# src/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Finance Dashboard API
  description: Backend system for a finance dashboard, supporting user management, financial record CRUD operations, aggregated analytics, and strict Role-Based Access Control (RBAC).
  version: 1.0.0
servers:
  - url: http://localhost:3000/api/v1
    description: Local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
...
  /users:
    get:
      summary: List all users (ADMIN only)
      tags: [Users]
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [ACTIVE, INACTIVE]
          description: Filter users by status
        - in: query
          name: page
          schema:
            type: integer
          description: Page number for pagination
        - in: query
          name: limit
          schema:
            type: integer
          description: Limit number of items per page
      responses:
        '200':
          description: A list of users
        '403':
          description: Forbidden (Requires ADMIN role)
    post:
      summary: Create a new user (ADMIN only)
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password, role]
              properties:
                name:
                  type: string
                email:
                  type: string
                password:
                  type: string
                role:
                  type: string
                  enum: [VIEWER, ANALYST, ADMIN]
      responses:
        '201':
          description: User created
        '403':
          description: Forbidden (Requires ADMIN role)
...
  /records:
    get:
      summary: List financial records
      tags: [Financial Records]
      parameters:
        - in: query
          name: type
          schema:
            type: string
            enum: [INCOME, EXPENSE]
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: from
          schema:
            type: string
            format: date
        - in: query
          name: to
          schema:
            type: string
            format: date
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: sort
          schema:
            type: string
            enum: [date, amount]
        - in: query
          name: order
          schema:
            type: string
            enum: [asc, desc]
      responses:
        '200':
          description: List of financial records
    post:
      summary: Create a financial record (ADMIN only)
      tags: [Financial Records]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount, type, category, date]
              properties:
                amount:
                  type: number
                type:
                  type: string
                  enum: [INCOME, EXPENSE]
                category:
                  type: string
                date:
                  type: string
                  format: date
                notes:
                  type: string
      responses:
        '201':
          description: Record created successfully
        '403':
          description: Forbidden (Requires ADMIN role)
...